fix: 글로벌 common 패키지 삭제
